% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_doc.R
\name{new_doc}
\alias{new_doc}
\title{Create a new quarto document based on a template}
\usage{
new_doc(
  outfile,
  template = "babraham_bioinformatics.qmd",
  author = "Simon Andrews",
  title = "Enter Title Here",
  theme = "theme_bw"
)
}
\arguments{
\item{outfile}{The name of the new document to create (needs .qmd suffix)}

\item{template}{The name of the template (defaults to standard Babraham template)}

\item{author}{Your name for the report}

\item{title}{The title of the document}

\item{theme}{Which ggplot theme to install (default theme_bw)}
}
\value{
Nothing, but a new document is created for you
}
\description{
Create a new quarto document based on a template
}
\examples{
new_doc("test.qmd", author="Jane Doe", title="My new analysis", theme="theme_dark")
}
